Fix curriculum profile permissions

+++ svelte/src/routes/cypress/reset-curriculum-profile.svelte
@@ -6,16 +6,23 @@
   let curriculumReset = false;
   let mounted = false;
   let db;
+  let deleted = false;
 
   async function removeCurriculumFeatures() {
+    deleted = true;
+    if (!$session.player.curriculumProfileId) {
+      curriculumReset = true;
+      return;
+    }
     let curriculumProfileRef = db
       .collection("curriculumProfile")
-      .doc($session.user.idTokenResult.claims.curriculumProfileId);
+      .doc($session.player.curriculumProfileId);
     try {
+      console.log("trying to delete token");
       await curriculumProfileRef.delete();
       let q = db
         .collection("revisions")
-        .where("authorId", "==", $session.user.idTokenResult.claims.curriculumProfileId);
+        .where("authorId", "==", $session.player.curriculumProfileId);
       const querySnapshot = await q.get();
       querySnapshot.forEach(async (snap) => {
         let revisionData = snap.data();
@@ -29,28 +36,44 @@
         }
         await goalRef.delete();
         await snap.ref.delete();
+        let playerRef = db.collection("players").doc($session.user.uid);
+        let playerSnap = await playerRef.get();
+        if (playerSnap.exists) {
+          let player = playerSnap.data();
+          delete player.curriculumProfileId;
+          playerRef.set(player);
+        }
       });
       curriculumReset = true;
     } catch (error) {
+      console.log("here?");
       console.error(error);
     }
   }
 
-  $: {
-    if (
-      mounted &&
-      ($session.environment === "cypress" ||
-        $session.environment === "development") &&
-      $session.user &&
-      $session.user.idTokenResult.claims.curriculumProfileId
-    ) {
-      removeCurriculumFeatures();
-    }
-  }
+  // $: {
+  //   if (
+  //     mounted &&
+  //     ($session.environment === "cypress" ||
+  //       $session.environment === "development") &&
+  //     $session.user &&
+  //     !deleted
+  //   ) {
+
+  //   }
+  // }
 
   onMount(async () => {
     firebase = await initFirebase($session.environment);
     db = await firebase.firestore();
+    if (
+      ($session.environment === "cypress" ||
+        $session.environment === "development") &&
+      $session.player &&
+      !deleted
+    ) {
+      await removeCurriculumFeatures();
+    }
     mounted = true;
   });
 </script>
